refactor: ai native contrib
