+++ packages/ai-native/src/browser/ai-editor.contribution.ts
@@ -1,84 +1,47 @@
 import debounce from 'lodash/debounce';
 
-import { Autowired, INJECTOR_TOKEN, Injectable, Injector } from '@opensumi/di';
-import { AINativeConfigService, IAIInlineChatService, PreferenceService } from '@opensumi/ide-core-browser';
+import { Autowired, Injectable } from '@opensumi/di';
+import { AINativeConfigService } from '@opensumi/ide-core-browser';
 import { IBrowserCtxMenu } from '@opensumi/ide-core-browser/lib/menu/next/renderer/ctxmenu/browser';
 import {
-  AIInlineChatContentWidgetId,
-  AINativeSettingSectionsId,
   AISerivceType,
-  AbortError,
-  CancelResponse,
   CancellationToken,
-  ChatResponse,
   ContributionProvider,
   Disposable,
-  ErrorResponse,
   Event,
   IAIReporter,
   IDisposable,
   IEventBus,
   ILogServiceClient,
   ILoggerManagerClient,
-  InlineChatFeatureRegistryToken,
-  MaybePromise,
-  ReplyResponse,
   Schemes,
   SupportLogNamespace,
   getErrorMessage,
-  runWhenIdle,
 } from '@opensumi/ide-core-common';
 import { DesignBrowserCtxMenuService } from '@opensumi/ide-design/lib/browser/override/menu.service';
-import { WorkbenchEditorService } from '@opensumi/ide-editor';
 import { EditorSelectionChangeEvent, IEditor, IEditorFeatureContribution } from '@opensumi/ide-editor/lib/browser';
 import { BrowserCodeEditor } from '@opensumi/ide-editor/lib/browser/editor-collection.service';
-import { WorkbenchEditorServiceImpl } from '@opensumi/ide-editor/lib/browser/workbench-editor.service';
 import * as monaco from '@opensumi/ide-monaco';
 import { monaco as monacoApi } from '@opensumi/ide-monaco/lib/browser/monaco-api';
-import { languageFeaturesService } from '@opensumi/ide-monaco/lib/browser/monaco-api/languages';
 import { MonacoTelemetryService } from '@opensumi/ide-monaco/lib/browser/telemetry.service';
-import { listenReadable } from '@opensumi/ide-utils/lib/stream';
 
-import { AI_DIFF_WIDGET_ID } from '../common';
-
-import { AINativeService } from './ai-native.service';
-import { AIInlineCompletionsProvider } from './inline-completions/completeProvider';
-import { AICompletionsService } from './inline-completions/service/ai-completions.service';
-import { LanguageParserService } from './languages/service';
-import { ICodeBlockInfo } from './languages/tree-sitter/language-facts/base';
-import { RenameSuggestionsService } from './rename/rename.service';
+import { CodeActionHandler } from './contrib/code-action/code-action.handler';
+import { AIInlineCompletionsProvider } from './contrib/inline-completions/completeProvider';
+import { AICompletionsService } from './contrib/inline-completions/service/ai-completions.service';
+import { RenameSuggestionsService } from './contrib/rename/rename.service';
 import { AINativeCoreContribution, IAIMiddleware } from './types';
-import { InlineChatController } from './widget/inline-chat/inline-chat-controller';
-import { InlineChatFeatureRegistry } from './widget/inline-chat/inline-chat.feature.registry';
-import { AIInlineChatService, EInlineChatStatus } from './widget/inline-chat/inline-chat.service';
-import { AIInlineContentWidget } from './widget/inline-chat/inline-content-widget';
-import { InlineDiffWidget } from './widget/inline-diff/inline-diff-widget';
+import { InlineChatEditorContribution } from './widget/inline-chat/inline-chat.contribution';
 
 @Injectable()
 export class AIEditorContribution extends Disposable implements IEditorFeatureContribution {
-  @Autowired(INJECTOR_TOKEN)
-  private readonly injector: Injector;
-
   @Autowired(AINativeConfigService)
   private readonly aiNativeConfigService: AINativeConfigService;
-
-  @Autowired(IAIInlineChatService)
-  private readonly aiInlineChatService: AIInlineChatService;
-
-  @Autowired(AINativeService)
-  private readonly aiNativeService: AINativeService;
 
   @Autowired(ILoggerManagerClient)
   private readonly loggerManagerClient: ILoggerManagerClient;
 
-  @Autowired(PreferenceService)
-  private readonly preferenceService: PreferenceService;
-
   @Autowired(IBrowserCtxMenu)
   private readonly ctxMenuRenderer: DesignBrowserCtxMenuService;
-
-  @Autowired(InlineChatFeatureRegistryToken)
-  private readonly inlineChatFeatureRegistry: InlineChatFeatureRegistry;
 
   @Autowired(IAIReporter)
   private readonly aiReporter: IAIReporter;
@@ -92,17 +55,17 @@ export class AIEditorContribution extends Disposable implements IEditorFeatureCo
   @Autowired(RenameSuggestionsService)
   private readonly renameSuggestionService: RenameSuggestionsService;
 
+  @Autowired(InlineChatEditorContribution)
+  private readonly inlineChatEditorContribution: InlineChatEditorContribution;
+
+  @Autowired(CodeActionHandler)
+  private readonly codeActionHandler: CodeActionHandler;
+
   @Autowired(AICompletionsService)
   private aiCompletionsService: AICompletionsService;
 
   @Autowired(IEventBus)
   private eventBus: IEventBus;
-
-  @Autowired(LanguageParserService)
-  private languageParserService: LanguageParserService;
-
-  @Autowired(WorkbenchEditorService)
-  private workbenchEditorService: WorkbenchEditorServiceImpl;
 
   @Autowired()
   private monacoTelemetryService: MonacoTelemetryService;
@@ -117,26 +80,8 @@ export class AIEditorContribution extends Disposable implements IEditorFeatureCo
     this.logger = this.loggerManagerClient.getLogger(SupportLogNamespace.Browser);
   }
 
-  private aiDiffWidget: InlineDiffWidget;
-  private aiInlineContentWidget: AIInlineContentWidget;
-  private aiInlineChatDisposed: Disposable = new Disposable();
-  private aiInlineChatOperationDisposed: Disposable = new Disposable();
-
   private modelSessionDisposable: Disposable;
   private initialized: boolean = false;
-
-  private disposeAllWidget() {
-    [
-      this.aiDiffWidget,
-      this.aiInlineContentWidget,
-      this.aiInlineChatDisposed,
-      this.aiInlineChatOperationDisposed,
-    ].forEach((widget) => {
-      widget?.dispose();
-    });
-
-    this.inlineChatInUsing = false;
-  }
 
   dispose(): void {
     super.dispose();
@@ -160,8 +105,9 @@ export class AIEditorContribution extends Disposable implements IEditorFeatureCo
 
         this.initialized = true;
         this.contributeInlineCompletionFeature(editor);
-        this.contributeInlineChatFeature(editor);
         this.registerLanguageFeatures(editor);
+
+        this.addDispose(this.inlineChatEditorContribution.contribute(editor));
       }),
     );
 
@@ -169,11 +115,9 @@ export class AIEditorContribution extends Disposable implements IEditorFeatureCo
 
     this.disposables.push(
       monacoEditor.onDidScrollChange(() => {
-        /**
-         * 其他的 ctxmenu 服务注册的菜单在 onHide 函数里会有其他逻辑处理，例如在 editor.context.ts 会在 hide 的时候 focus 编辑器，影响使用
-         */
-        this.ctxMenuRenderer.onHide = undefined;
-        this.ctxMenuRenderer.hide(true);
+        if (this.ctxMenuRenderer.visible) {
+          this.ctxMenuRenderer.hide(true);
+        }
       }),
     );
 
@@ -227,90 +171,10 @@ export class AIEditorContribution extends Disposable implements IEditorFeatureCo
       }),
       monacoEditor.onWillChangeModel(() => {
         this.aiCompletionsService.hideStatusBarItem();
-        this.disposeAllWidget();
       }),
       monacoEditor.onDidBlurEditorText(() => {
         this.aiCompletionsService.hideStatusBarItem();
         this.aiCompletionsService.setVisibleCompletion(false);
-      }),
-    );
-  }
-
-  protected contributeInlineChatFeature(editor: IEditor): void {
-    const { monacoEditor } = editor;
-
-    this.disposables.push(
-      this.aiNativeService.onInlineChatVisible((value: boolean) => {
-        if (value) {
-          this.showInlineChat(editor);
-        } else {
-          this.aiNativeService.cancelToken();
-          this.disposeAllWidget();
-        }
-      }),
-      // 通过 code actions 来透出我们 inline chat 的功能
-      this.inlineChatFeatureRegistry.onCodeActionRun(({ id, range }) => {
-        const currentEditor = this.workbenchEditorService.currentEditor;
-
-        if (currentEditor?.currentUri !== editor.currentUri) {
-          return;
-        }
-
-        monacoEditor.setSelection(range);
-        this.showInlineChat(editor);
-        if (this.aiInlineContentWidget) {
-          this.aiInlineContentWidget.clickActionId(id, 'codeAction');
-        }
-      }),
-    );
-
-    let needShowInlineChat = false;
-    this.disposables.push(
-      monacoEditor.onMouseDown(() => {
-        needShowInlineChat = false;
-      }),
-      monacoEditor.onMouseUp((event) => {
-        const target = event.target;
-        const detail = (target as any).detail;
-        if (detail && typeof detail === 'string' && detail === AIInlineChatContentWidgetId) {
-          needShowInlineChat = false;
-        } else {
-          needShowInlineChat = true;
-        }
-      }),
-    );
-
-    let prefInlineChatAutoVisible = this.preferenceService.getValid(
-      AINativeSettingSectionsId.INLINE_CHAT_AUTO_VISIBLE,
-      true,
-    );
-    this.disposables.push(
-      this.preferenceService.onSpecificPreferenceChange(
-        AINativeSettingSectionsId.INLINE_CHAT_AUTO_VISIBLE,
-        ({ newValue }) => {
-          prefInlineChatAutoVisible = newValue;
-        },
-      ),
-    );
-
-    this.disposables.push(
-      Event.debounce(
-        Event.any<any>(monacoEditor.onDidChangeCursorSelection, monacoEditor.onMouseUp),
-        (_, e) => e,
-        100,
-      )(() => {
-        if (!prefInlineChatAutoVisible || !needShowInlineChat) {
-          return;
-        }
-
-        if (
-          this.aiInlineChatService.status !== EInlineChatStatus.READY &&
-          this.aiInlineChatService.status !== EInlineChatStatus.ERROR
-        ) {
-          return;
-        }
-
-        this.showInlineChat(editor);
       }),
     );
   }
@@ -321,346 +185,6 @@ export class AIEditorContribution extends Disposable implements IEditorFeatureCo
     }
 
     return false;
-  }
-
-  protected inlineChatInUsing = false;
-  protected async showInlineChat(editor: IEditor): Promise<void> {
-    if (!this.aiNativeConfigService.capabilities.supportsInlineChat) {
-      return;
-    }
-    if (this.inlineChatInUsing) {
-      return;
-    }
-
-    this.inlineChatInUsing = true;
-
-    this.disposeAllWidget();
-
-    const { monacoEditor } = editor;
-
-    const selection = monacoEditor.getSelection();
-
-    if (!selection || selection.isEmpty()) {
-      this.disposeAllWidget();
-      return;
-    }
-
-    this.aiInlineChatDisposed.addDispose(this.aiInlineChatService.launchChatStatus(EInlineChatStatus.READY));
-
-    this.aiInlineContentWidget = this.injector.get(AIInlineContentWidget, [monacoEditor]);
-
-    this.aiInlineContentWidget.show({
-      selection,
-    });
-
-    this.aiInlineChatDisposed.addDispose(
-      this.aiInlineContentWidget.onActionClick((action) => {
-        this.runInlineChatAction(action, monacoEditor);
-      }),
-    );
-  }
-
-  private async runInlineChatAction(
-    {
-      actionId: id,
-      source,
-    }: {
-      actionId: string;
-      source: string;
-    },
-    monacoEditor: monaco.ICodeEditor,
-  ) {
-    const handler = this.inlineChatFeatureRegistry.getEditorHandler(id);
-    const action = this.inlineChatFeatureRegistry.getAction(id);
-    if (!handler || !action) {
-      return;
-    }
-
-    const selection = monacoEditor.getSelection();
-    if (!selection) {
-      this.logger.error('No selection found, aborting inline chat action.');
-      return;
-    }
-
-    const { execute, providerDiffPreviewStrategy } = handler;
-
-    if (execute) {
-      await execute(monacoEditor);
-      this.disposeAllWidget();
-    }
-
-    if (providerDiffPreviewStrategy) {
-      const crossSelection = selection
-        .setStartPosition(selection.startLineNumber, 1)
-        .setEndPosition(selection.endLineNumber, Number.MAX_SAFE_INTEGER);
-
-      const relationId = this.aiReporter.start(action.name, {
-        message: action.name,
-        type: AISerivceType.InlineChat,
-        source,
-        runByCodeAction: source === 'codeAction',
-      });
-
-      await this.handleDiffPreviewStrategy(
-        monacoEditor,
-        providerDiffPreviewStrategy,
-        crossSelection,
-        relationId,
-        false,
-      );
-
-      this.aiInlineChatDisposed.addDispose([
-        this.aiInlineChatService.onDiscard(() => {
-          this.aiReporter.end(relationId, { message: 'discard', success: true, isDrop: true });
-          this.disposeAllWidget();
-        }),
-        this.aiInlineChatService.onRegenerate(async () => {
-          await this.handleDiffPreviewStrategy(
-            monacoEditor,
-            providerDiffPreviewStrategy,
-            crossSelection,
-            relationId,
-            true,
-          );
-        }),
-      ]);
-    }
-  }
-
-  private async handleDiffPreviewStrategy(
-    monacoEditor: monaco.ICodeEditor,
-    strategy: (
-      editor: monaco.ICodeEditor,
-      cancelToken: CancellationToken,
-    ) => MaybePromise<ChatResponse | InlineChatController>,
-    crossSelection: monaco.Selection,
-    relationId: string,
-    isRetry: boolean,
-  ): Promise<void> {
-    const model = monacoEditor.getModel();
-
-    this.aiDiffWidget?.dispose();
-    this.aiInlineChatOperationDisposed.dispose();
-    this.aiInlineChatDisposed.addDispose(this.aiInlineChatService.launchChatStatus(EInlineChatStatus.THINKING));
-
-    const startTime = Date.now();
-
-    if (this.aiNativeService.cancelIndicator.token.isCancellationRequested) {
-      this.convertInlineChatStatus(EInlineChatStatus.READY, {
-        relationId,
-        message: 'abort',
-        startTime,
-        isRetry,
-        isStop: true,
-      });
-      return;
-    }
-
-    const response = await strategy(monacoEditor, this.aiNativeService.cancelIndicator.token);
-
-    this.visibleDiffWidget(
-      monacoEditor,
-      { crossSelection, chatResponse: response },
-      { relationId, startTime, isRetry },
-    );
-
-    this.aiInlineChatOperationDisposed.addDispose([
-      this.aiInlineChatService.onAccept(() => {
-        this.aiReporter.end(relationId, { message: 'accept', success: true, isReceive: true });
-        const newValue = this.aiDiffWidget?.getModifiedModel()?.getValue() || '';
-
-        monacoEditor.getModel()?.pushEditOperations(null, [{ range: crossSelection, text: newValue }], () => null);
-        runWhenIdle(() => {
-          this.disposeAllWidget();
-        });
-      }),
-      this.aiInlineChatService.onThumbs((isLike: boolean) => {
-        this.aiReporter.end(relationId, { isLike });
-      }),
-      this.aiDiffWidget.onMaxLineCount((count) => {
-        requestAnimationFrame(() => {
-          if (crossSelection.endLineNumber === model!.getLineCount()) {
-            // 如果用户是选中了最后一行，直接显示在最后一行
-            const lineHeight = monacoEditor.getOption(monacoApi.editor.EditorOption.lineHeight);
-            this.aiInlineContentWidget.offsetTop(lineHeight * count + 12);
-          }
-        });
-      }),
-    ]);
-  }
-
-  private formatAnswer(answer: string, crossCode: string): string {
-    const leadingWhitespaceMatch = crossCode.match(/^\s*/);
-    const indent = leadingWhitespaceMatch ? leadingWhitespaceMatch[0] : '  ';
-    return answer
-      .split('\n')
-      .map((line) => `${indent}${line}`)
-      .join('\n');
-  }
-
-  private convertInlineChatStatus(
-    status: EInlineChatStatus,
-    reportInfo: {
-      relationId: string;
-      message: string;
-      startTime: number;
-      isRetry?: boolean;
-      isStop?: boolean;
-    },
-  ): void {
-    const { relationId, message, startTime, isRetry, isStop } = reportInfo;
-
-    this.aiInlineChatDisposed.addDispose(this.aiInlineChatService.launchChatStatus(status));
-    this.aiReporter.end(relationId, {
-      message,
-      success: status !== EInlineChatStatus.ERROR,
-      replytime: Date.now() - startTime,
-      isStop,
-      isRetry,
-    });
-  }
-
-  private visibleDiffWidget(
-    monacoEditor: monaco.ICodeEditor,
-    options: {
-      crossSelection: monaco.Selection;
-      chatResponse?: ChatResponse | InlineChatController;
-    },
-    reportInfo: {
-      relationId: string;
-      startTime: number;
-      isRetry: boolean;
-    },
-  ): void {
-    const { crossSelection, chatResponse } = options;
-    const { relationId, startTime, isRetry } = reportInfo;
-
-    this.aiDiffWidget = this.injector.get(InlineDiffWidget, [
-      AI_DIFF_WIDGET_ID,
-      {
-        editor: monacoEditor,
-        selection: crossSelection,
-      },
-    ]);
-    this.aiDiffWidget.create();
-    this.aiDiffWidget.showByLine(
-      crossSelection.startLineNumber - 1,
-      crossSelection.endLineNumber - crossSelection.startLineNumber + 2,
-    );
-
-    if (InlineChatController.is(chatResponse)) {
-      const controller = chatResponse as InlineChatController;
-
-      this.aiInlineChatOperationDisposed.addDispose(
-        this.aiDiffWidget.onReady(() => {
-          const modifiedModel = this.aiDiffWidget.getModifiedModel();
-          if (!modifiedModel) {
-            return;
-          }
-
-          let isAbort = false;
-
-          this.aiInlineChatOperationDisposed.addDispose([
-            controller.onData((data) => {
-              if (ReplyResponse.is(data)) {
-                isAbort = false;
-                const { message } = data;
-
-                const lastLine = modifiedModel.getLineCount();
-                const lastColumn = modifiedModel.getLineMaxColumn(lastLine);
-
-                const range = new monaco.Range(lastLine, lastColumn, lastLine, lastColumn);
-
-                const edit = {
-                  range,
-                  text: message || '',
-                };
-                modifiedModel.pushEditOperations(null, [edit], () => null);
-                this.aiDiffWidget.layout();
-              }
-            }),
-            controller.onError((error) => {
-              this.convertInlineChatStatus(EInlineChatStatus.ERROR, {
-                relationId,
-                message: error.message || '',
-                startTime,
-                isRetry,
-              });
-            }),
-            controller.onAbort(() => {
-              this.convertInlineChatStatus(EInlineChatStatus.READY, {
-                relationId,
-                message: 'abort',
-                startTime,
-                isRetry,
-                isStop: true,
-              });
-            }),
-            controller.onEnd(() => {
-              this.convertInlineChatStatus(EInlineChatStatus.DONE, {
-                relationId,
-                message: '',
-                startTime,
-                isRetry,
-              });
-            }),
-          ]);
-        }),
-      );
-    } else {
-      const model = monacoEditor.getModel();
-      const crossCode = model!.getValueInRange(crossSelection);
-
-      if (this.aiInlineChatDisposed.disposed || CancelResponse.is(chatResponse)) {
-        this.convertInlineChatStatus(EInlineChatStatus.READY, {
-          relationId,
-          message: (chatResponse as CancelResponse).message || '',
-          startTime,
-          isRetry,
-          isStop: true,
-        });
-        return;
-      }
-
-      if (ErrorResponse.is(chatResponse)) {
-        this.convertInlineChatStatus(EInlineChatStatus.ERROR, {
-          relationId,
-          message: (chatResponse as ErrorResponse).message || '',
-          startTime,
-          isRetry,
-        });
-        return;
-      }
-
-      this.convertInlineChatStatus(EInlineChatStatus.DONE, {
-        relationId,
-        message: '',
-        startTime,
-        isRetry,
-      });
-
-      let answer = (chatResponse as ReplyResponse).message;
-      answer = this.formatAnswer(answer, crossCode);
-
-      this.aiInlineChatOperationDisposed.addDispose(
-        this.aiDiffWidget.onReady(() => {
-          const modifiedModel = this.aiDiffWidget.getModifiedModel();
-          if (!modifiedModel) {
-            return;
-          }
-
-          modifiedModel.setValue(answer);
-        }),
-      );
-    }
-
-    this.aiInlineContentWidget?.setOptions({
-      position: {
-        lineNumber: crossSelection.endLineNumber + 1,
-        column: 1,
-      },
-    });
-    this.aiInlineContentWidget?.layoutContentWidget();
   }
 
   private async registerLanguageFeatures(editor: IEditor): Promise<void> {
@@ -735,7 +259,7 @@ export class AIEditorContribution extends Disposable implements IEditorFeatureCo
       }
 
       if (this.aiNativeConfigService.capabilities.supportsInlineChat) {
-        this.modelSessionDisposable.addDispose(this.contributeCodeActionFeature(languageId, editor));
+        this.modelSessionDisposable.addDispose(this.codeActionHandler.registerCodeActionFeature(languageId, editor));
       }
     };
 
@@ -813,132 +337,5 @@ export class AIEditorContribution extends Disposable implements IEditorFeatureCo
     ]);
 
     return disposable;
-  }
-
-  protected contributeCodeActionFeature(languageId: string, editor: IEditor): IDisposable {
-    const disposable = new Disposable();
-
-    let prefInlineChatActionEnabled = this.preferenceService.getValid(
-      AINativeSettingSectionsId.INLINE_CHAT_CODE_ACTION_ENABLED,
-      true,
-    );
-
-    if (!prefInlineChatActionEnabled) {
-      return disposable;
-    }
-
-    const { monacoEditor } = editor;
-    const { languageParserService, inlineChatFeatureRegistry, shouldAbortRequest } = this;
-
-    let codeActionDispose: IDisposable | undefined;
-
-    disposable.addDispose(
-      this.preferenceService.onSpecificPreferenceChange(
-        AINativeSettingSectionsId.INLINE_CHAT_CODE_ACTION_ENABLED,
-        ({ newValue }) => {
-          prefInlineChatActionEnabled = newValue;
-          if (newValue) {
-            register();
-          } else {
-            if (codeActionDispose) {
-              codeActionDispose.dispose();
-              codeActionDispose = undefined;
-            }
-          }
-        },
-      ),
-    );
-
-    register();
-
-    return disposable;
-
-    function register() {
-      if (codeActionDispose) {
-        codeActionDispose.dispose();
-        codeActionDispose = undefined;
-      }
-
-      codeActionDispose = languageFeaturesService.codeActionProvider.register(languageId, {
-        provideCodeActions: async (model) => {
-          if (shouldAbortRequest(model)) {
-            return;
-          }
-
-          if (!prefInlineChatActionEnabled) {
-            return;
-          }
-
-          const parser = languageParserService.createParser(languageId);
-          if (!parser) {
-            return;
-          }
-          const actions = inlineChatFeatureRegistry.getCodeActions();
-          if (!actions || actions.length === 0) {
-            return;
-          }
-
-          const cursorPosition = monacoEditor.getPosition();
-          if (!cursorPosition) {
-            return;
-          }
-
-          function constructCodeActions(info: ICodeBlockInfo) {
-            return {
-              actions: actions.map((v) => {
-                const command = {} as monaco.Command;
-                if (v.command) {
-                  command.id = v.command.id;
-                  command.arguments = [info.range];
-                }
-
-                let title = v.title;
-
-                switch (info.infoCategory) {
-                  case 'function': {
-                    title = title + ` for Function: ${info.name}`;
-                  }
-                }
-
-                return {
-                  ...v,
-                  title,
-                  ranges: [info.range],
-                  command,
-                };
-              }) as monaco.CodeAction[],
-              dispose() {},
-            };
-          }
-
-          const info = await parser.provideCodeBlockInfo(model, cursorPosition);
-          if (info) {
-            return constructCodeActions(info);
-          }
-
-          // check current line is empty
-          const currentLineLength = model.getLineLength(cursorPosition.lineNumber);
-          if (currentLineLength !== 0) {
-            return;
-          }
-
-          // 获取视窗范围内的代码块
-          const ranges = monacoEditor.getVisibleRanges();
-          if (ranges.length === 0) {
-            return;
-          }
-
-          // 查找从当前行至视窗最后一行的代码块中是否包含函数
-          const newRange = new monaco.Range(cursorPosition.lineNumber, 0, ranges[0].endLineNumber + 1, 0);
-
-          const rangeInfo = await parser.provideCodeBlockInfoInRange(model, newRange);
-          if (rangeInfo) {
-            return constructCodeActions(rangeInfo);
-          }
-        },
-      });
-
-      disposable.addDispose(codeActionDispose);
-    }
   }
 }
