+++ packages/ai-native/src/browser/contrib/inline-completions/model/competionModel.ts
