+++ packages/ai-native/src/browser/contrib/inline-completions/prompt/const.ts
