+++ packages/ai-native/src/browser/contrib/inline-completions/prompt/importedFiles.ts
@@ -4,7 +4,7 @@ import { IFileServiceClient } from '@opensumi/ide-file-service';
 import { Path } from '@opensumi/ide-utils/lib/path';
 import { IWorkspaceService } from '@opensumi/ide-workspace';
 
-import { LanguageParserService } from '../../languages/service';
+import { LanguageParserService } from '../../../languages/service';
 import { ICompletionContext, ImportedFileOptions, ResourceDocument } from '../types';
 
 import { LANGUAGE_TO_SUFFIX } from './languages';
