+++ packages/ai-native/src/browser/contrib/inline-completions/promptCache.ts
