+++ packages/ai-native/src/browser/contrib/inline-completions/provider.ts
