+++ packages/ai-native/src/browser/contrib/inline-completions/types.ts
