+++ packages/ai-native/src/browser/contrib/interface-navigation/interface-navigation.module.less
