+++ packages/ai-native/src/browser/contrib/merge-conflict/index.ts
