+++ packages/ai-native/src/browser/contrib/terminal/component/terminal-command-suggest-controller.module.less
