+++ packages/ai-native/src/browser/contrib/terminal/component/terminal-inline-chat-controller.tsx
