+++ packages/ai-native/src/browser/contrib/terminal/index.ts
