+++ packages/ai-native/src/browser/contrib/terminal/matcher/base.ts
