+++ packages/ai-native/src/browser/contrib/terminal/matcher/index.ts
