+++ packages/ai-native/src/browser/contrib/terminal/matcher/node.ts
