+++ packages/ai-native/src/browser/contrib/terminal/matcher/npm.ts
