+++ packages/ai-native/src/browser/contrib/terminal/matcher/shell.ts
