+++ packages/ai-native/src/browser/contrib/terminal/terminal-ai.contributon.ts
