+++ packages/ai-native/src/browser/contrib/terminal/utils/ansi-parser.ts
