+++ packages/ai-native/src/browser/widget/light-bulb/index.ts
@@ -7,7 +7,7 @@ import {
   LightBulbWidget,
 } from '@opensumi/monaco-editor-core/esm/vs/editor/contrib/codeAction/browser/lightBulbWidget';
 
-export class OpenSumiLightBulbWidget extends LightBulbWidget {
+export class SumiLightBulbWidget extends LightBulbWidget {
   protected override _updateLightBulbTitleAndIcon(): void {
     this._domNode.classList.remove(...this._iconClasses);
     this._iconClasses = [];
